requirements.txt
streamlit
pandas
python-bcb
plotly
datetime